Made data._id write to disk, started new filtering
In libs/model.js, moved the code:
data._id = autoinc
from line 59 to 57. This includes it in the data that is wrote to disk.
Previously, when the program ended, the _id value did not persist since
it only existed in the copy of the data that existed in memory.
Therefore, a query in a previous iteration of the program that relied on
searching by id would fail.

Added libs/filter.js that adds in model querying similar to loopback
model querying. Currently testing changes, will provide proper
documentation afterwards.