Documentation and testing of filters completed.
Summary of better-filter feature branch changes:

Incremented version to v0.1.1

docs/test.js contains a full testing example for reference purposes.

Added libs/filter.js containing filter/querying operations to make the
database 'smarter'. See the Filtering section of the readme for
information.

Modified libs/model.js to use libs/filter.js to find what rows of the
database the find/update/remove method to operate on.
It is important to note that the _id field is now included in the data
written to storage, which is necessary for database operations relying
on a unique identifier to persist.

Added information about filtering to the readme and updated the example
to reflect the new information.